A bit of functional code golfing. I'm not sure if it's more or less readable heh

+++ AdventOfCode2023/Day9/Tests.fs
@@ -5,7 +5,7 @@ open Xunit
 [<Theory>]
 [<InlineData("Day9/testInput.txt", 114)>]
 [<InlineData("Day9/input.txt", 1930746032)>]
-let ``The sum of the future extrapolated values`` (filename: string, expected: int64) =
+let ``The sum of the future extrapolated values`` (filename: string, expected: int) =
   let result =
     filename
     |> OASISReport.parse
@@ -18,7 +18,7 @@ let ``The sum of the future extrapolated values`` (filename: string, expected: i
 [<Theory>]
 [<InlineData("Day9/testInput.txt", 2)>]
 [<InlineData("Day9/input.txt", 1154)>]
-let ``The sum of the past extrapolated values`` (filename: string, expected: int64) =
+let ``The sum of the past extrapolated values`` (filename: string, expected: int) =
   let result =
     filename
     |> OASISReport.parse
